backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
geoalchemy2==0.15.2
python-dotenv==1.0.1
pydantic-settings==2.5.2
httpx==0.27.2
google-genai>=1.0.0
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
slowapi>=0.1.9
cryptography>=42.0.0
